% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff.r
\name{ff}
\alias{ff}
\title{generate forest patches with the forest factory}
\usage{
ff(data = "temperate_general")
}
\arguments{
\item{data}{characterstring - A forest parameterset; default "temperate_general", Furhter available dataset: "amazon_general"}
}
\description{
generate various forest patches with the forest factory algorihtm
}
